refactor(TableSection): move component to Table folder
