+++ packages/core/src/Table/TableSection/TableSection.styles.tsx
@@ -1,12 +1,12 @@
 import { createClasses } from "@hitachivantara/uikit-react-utils";
 import { theme } from "@hitachivantara/uikit-styles";
 
-import { bulkActionsClasses } from "../BulkActions";
-import { paginationClasses } from "../Pagination";
-import { tableCellClasses } from "../Table/TableCell";
-import { tableContainerClasses } from "../Table/TableContainer";
-import { tableHeaderClasses } from "../Table/TableHeader";
-import { tableRowClasses } from "../Table/TableRow";
+import { bulkActionsClasses } from "../../BulkActions";
+import { paginationClasses } from "../../Pagination";
+import { tableCellClasses } from "../TableCell";
+import { tableContainerClasses } from "../TableContainer";
+import { tableHeaderClasses } from "../TableHeader";
+import { tableRowClasses } from "../TableRow";
 
 export const { staticClasses, useClasses } = createClasses("HvTableSection", {
   root: {},
